+++ pkg/bpmn_engine/expressions.go
@@ -13,13 +13,13 @@ func evaluateExpression(expression string, variableContext map[string]interface{
 	return expr.Eval(expression, variableContext)
 }
 
-func evaluateLocalVariables(varHolder var_holder.VariableHolder, mappings []extensions.TIoMapping) error {
+func evaluateLocalVariables(varHolder *var_holder.VariableHolder, mappings []extensions.TIoMapping) error {
 	return mapVariables(varHolder, mappings, func(key string, value interface{}) {
 		varHolder.SetVariable(key, value)
 	})
 }
 
-func propagateProcessInstanceVariables(varHolder var_holder.VariableHolder, mappings []extensions.TIoMapping) error {
+func propagateProcessInstanceVariables(varHolder *var_holder.VariableHolder, mappings []extensions.TIoMapping) error {
 	if len(mappings) == 0 {
 		for k, v := range varHolder.Variables() {
 			varHolder.PropagateVariable(k, v)
@@ -30,7 +30,7 @@ func propagateProcessInstanceVariables(varHolder var_holder.VariableHolder, mapp
 	})
 }
 
-func mapVariables(varHolder var_holder.VariableHolder, mappings []extensions.TIoMapping, setVarFunc func(key string, value interface{})) error {
+func mapVariables(varHolder *var_holder.VariableHolder, mappings []extensions.TIoMapping, setVarFunc func(key string, value interface{})) error {
 	for _, mapping := range mappings {
 		evalResult, err := evaluateExpression(mapping.Source, varHolder.Variables())
 		if err != nil {